+++ DTCD-StyleSystem/src/stories/BaseSwitch.stories.js
@@ -2,6 +2,20 @@ import BaseSwitchDoc from './docs/BaseSwitchDoc.mdx';
 
 export default {
   title: 'Example/BaseComponents/BaseSwitch',
+  argTypes: {
+    defaultSlot: { 
+      type: 'string',
+    },
+    type: { 
+      control: {
+        type: 'select',
+      },
+      options: [
+        'checkbox',
+        'radio',
+      ],
+    },
+  },
   parameters: {
     docs: {
       page: BaseSwitchDoc,
@@ -11,19 +25,21 @@ export default {
 
 const Template = (args) => {
   const {
+    type,
     disabled,
     checked,
-    value,
     label,
+    value,
     labelSlot,
   } = args;
 
   const switchInput = document.createElement('base-switch');
-
+  
   switchInput.checked = checked;
+  switchInput.type = type;
   switchInput.disabled = disabled;
-  switchInput.value = value;
   switchInput.label = label;
+  switchInput.value = value;
   
   switchInput.innerHTML += labelSlot;
 
@@ -36,5 +52,6 @@ DefaultSwitch.args = {
   disabled: false,
   checked: false,
   value: '',
+  type: 'checkbox',
   labelSlot: '<span slot="label">Label slot</span>',
 };