+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - how `invalid` property works. Edited BaseInput, BaseSelect, BaseTextarea
 - BaseSelect rendering so that it doesn't get cut off
 - definition of custom html-elements in Storybook (added try-catch statement)
+- joined JS-code of BaseCheckbox, BaseRadio and BaseSwitc
 
 ## [0.14.0]